javascript/apps/dash/backend/src/grpc/__tests__/builders.test.ts
import {fileFromObject, fileInfoFromObject} from '@dash-backend/grpc/builders';

describe('grpc/pb', () => {
  it('should create File from an object', () => {
    const file = fileFromObject({
      commitId: '1234567890',
      path: '/assets',
      repoName: 'neato',
    });

    expect(file.getCommit()?.getId()).toBe('1234567890');
    expect(file.getCommit()?.getRepo()?.getName()).toBe('neato');
    expect(file.getPath()).toBe('/assets');
  });

  it('should create File from an object with defaults', () => {
    const file = fileFromObject({
      repoName: 'neato',
    });

    expect(file.getCommit()?.getId()).toBe('master');
    expect(file.getCommit()?.getRepo()?.getName()).toBe('neato');
    expect(file.getPath()).toBe('/');
  });

  it('should create FileInfo from an object', () => {
    const fileInfo = fileInfoFromObject({
      file: {
        commitId: '1234567890',
        path: '/assets',
        repoName: 'neato',
      },
      fileType: 2,
      hash: 'abcde12345',
      sizeBytes: 123,
    });

    expect(fileInfo.getFile()?.getCommit()?.getId()).toBe('1234567890');
    expect(fileInfo.getFile()?.getCommit()?.getRepo()?.getName()).toBe('neato');
    expect(fileInfo.getFile()?.getPath()).toBe('/assets');
    expect(fileInfo.getFileType()).toBe(2);
    expect(fileInfo.getHash()).toBe('abcde12345');
    expect(fileInfo.getSizeBytes()).toBe(123);
  });
});
